# docs/v1/swagger.yaml
definitions:
  dto.CategoryRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      suggested_percentage:
        type: integer
    type: object
  dto.CategoryResponse:
    properties:
      color:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      suggested_percentage:
        type: integer
    type: object
...
  dto.LoginRequest:
    properties:
      identifier:
        description: username or email
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
info:
  contact: {}
  title: API Frog-Go
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Autentica o usuário pelo username **ou** email e senha, retornando
        um token JWT
      parameters:
      - description: Credenciais de login (identifier = username or email)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
      summary: Login
      tags:
      - Auth
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: Lista todas as categorias aplicando filtros e paginação
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Limite por página
        in: query
        name: limit
        type: integer
      - description: 'Campo de ordenação (ex: name)'
        in: query
        name: order_by
        type: string
      - description: Ordem (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista categorias com filtros e paginação
      tags:
      - Categorias
    post:
      consumes:
      - application/json
      description: Cria uma nova categoria com os dados fornecidos no corpo da requisição
      parameters:
      - description: Dados da categoria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
      security:
      - BearerAuth: []
      summary: Cria uma nova categoria
      tags:
      - Categorias
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui uma categoria com base no ID fornecido
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Sem conteúdo
      security:
      - BearerAuth: []
      summary: Remove uma categoria
      tags:
      - Categorias
    get:
      consumes:
      - application/json
      description: Retorna os dados de uma categoria com base no ID fornecido
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
      security:
      - BearerAuth: []
      summary: Busca uma categoria por ID
      tags:
      - Categorias
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma categoria com base no ID fornecido e nos
        dados enviados no corpo da requisição
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados da categoria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
      security:
      - BearerAuth: []
      summary: Atualiza uma categoria existente
      tags:
      - Categorias
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT no formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"